ai-service/requirements.txt
# ── Web framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Embeddings ───────────────────────────────────────────────
sentence-transformers==3.0.0
torch==2.2.2               # CPU build; change to torch+cu118 for GPU

# ── Vector store ─────────────────────────────────────────────
faiss-cpu==1.8.0           # faiss-gpu if you have CUDA

# ── LLM — Ollama client ──────────────────────────────────────
ollama==0.2.1

# ── Data / validation ─────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
numpy==1.26.4

# ── Persistence (FAISS index snapshots to disk) ──────────────
# Built-in pickle used; no extra dep needed

# ── Logging / utilities ──────────────────────────────────────
loguru==0.7.2
python-dotenv==1.0.1

# ── Testing ──────────────────────────────────────────────────
pytest==8.2.0
httpx==0.27.0              # async test client for FastAPI
